requirements.txt
# Core dependencies
dash==2.14.2
dash-bootstrap-components==1.5.0
pandas==2.1.4
plotly==5.19.0
SQLAlchemy==2.0.25
dash-draggable==0.1.2
python-dotenv==1.0.1
APScheduler==3.10.4
Flask-APScheduler==1.13.1
pytz==2024.1
tzlocal==5.2

# Development and testing
pytest==8.1.1
pytest-cov==4.1.0
pytest-mock==3.14.0
coverage==7.4.4

# Code quality
black==24.3.0
flake8==7.0.0
isort==5.13.2
mypy==1.8.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.2
